Rename gui folder to api
